Move move-native modules to files
